Created shared file for test and kafkagen to use to generate data through rest proxy
